Refactor ProtelisRuntimeException
