+++ protelis/protelis-interpreter/src/main/java/org/protelis/lang/interpreter/util/ProtelisRuntimeException.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package org.protelis.lang;
+package org.protelis.lang.interpreter.util;
 
 import java.io.PrintStream;
 import java.util.Deque;
